requirements.txt
absl-py==0.8.1
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.3.0
billiard==3.3.0.23
cachetools==3.1.1
celery==3.1.26.post2
celery-with-redis==3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
cryptography==2.7
cycler==0.10.0
Cython==0.29.13
dask==2.8.0
DateTime==4.3
decorator==4.4.1
defusedxml==0.6.0
Django==2.1.9
django-celery==3.3.1
django-cors-headers==3.0.2
django-crontab==0.7.1
django-easy-pdf==0.1.1
django-filter==2.1.0
django-haystack==2.8.1
django-mssql==1.8
django-object-tool==0.0.8
django-pdfkit==0.3.1
django-pyodbc==1.1.3
django-pyodbc-azure==2.1.0.0
django-pytds==1.5
django-sqlserver==1.11
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-xml==1.4.0
drf-extensions==0.5.0
drf-writable-nested==0.5.1
fasttext==0.9.1
filterpy==1.4.1
Flask==1.0.2
Flask-WTF==0.14.2
future==0.18.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
grpcio==1.25.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
imgviz==0.10.5
importlib-metadata==0.23
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10
jsonfield==2.0.2
kiwisolver==1.1.0
kombu==3.0.37
labelme==3.21.1
llvmlite==0.30.0
lxml==4.3.3
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.1.1
more-itertools==7.2.0
munch==2.5.0
Naked==0.1.31
networkx==2.4
nibabel==2.5.0
numba==0.38.1
numpy==1.15.0
oauthlib==3.1.0
opencv-contrib-python==4.1.1.26
opencv-python==4.1.0.25
optionaldict==0.1.1
packaging==19.2
pandas==0.25.1
pdfkit==0.6.0
Pillow==5.3.0
pluggy==0.13.0
pretrainedmodels==0.7.4
protobuf==3.10.0
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pybind11==2.3.0
pycocotools==2.0.0
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.8.2
pyDes==2.0.1
Pygments==2.4.2
PyJWT==1.7.1
pymssql==2.1.4
PyMySQL==0.9.3
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.2
PyPDF2==1.26.0
PyQt5==5.14.1
PyQt5-sip==12.7.0
pytest==5.2.4
python-dateutil==2.8.0
python-tds==1.9.1
pytz==2019.1
PyWavelets==1.1.1
PyYAML==5.1.1
pyzmq==18.1.0
QtPy==1.9.0
redis==2.10.6
reportlab==3.5.23
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==1.3.1
sentry-sdk==0.12.1
shellescape==3.4.1
six==1.10.0
sklearn==0.0
sqlparse==0.3.0
tb-nightly==2.1.0a20191117
termcolor==1.1.0
toolz==0.10.0
tor==1.0.0
torch==1.2.0
torchfile==0.1.0
torchvision==0.4.0
tornado==6.0.3
tqdm==4.35.0
uritemplate==3.0.0
urllib3==1.25.3
visdom==0.1.8.8

websocket-client==0.56.0
wechat==0.4.16
wechat-django==0.3.3
wechat-python-sdk==0.6.1
wechatpy==1.8.3
